[WPE] Follow-up 294148@main, make 'fps' a float instead of double https://bugs.webkit.org/show_bug.cgi?id=278460
Reviewed by Carlos Garcia Campos.

Changeset 294148@main made use of the data type
std::atomic<std::optional<double>>. This data type is 9-bytes long,
which it seems can only be handled in 32-bit systems if libatomic is
linked.

It does not seem variable 'fps' needs to be a double, and could be a
float instead. Reducing the data size fixes the build without requiring
libatomic.

* Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/ThreadedCompositor.h:
(WebKit::ThreadedCompositor::fps const):

Canonical link: https://commits.webkit.org/294245@main

+++ Source/WebKit/WebProcess/WebPage/CoordinatedGraphics/ThreadedCompositor.h
@@ -102,7 +102,7 @@ public:
 
     bool isActive() const;
 
-    std::optional<double> fps() const { return m_fpsCounter.fps.load(); };
+    std::optional<float> fps() const { return m_fpsCounter.fps.load(); };
 
 #if ENABLE(DAMAGE_TRACKING)
     void setDamagePropagation(WebCore::Damage::Propagation);
@@ -161,7 +161,7 @@ private:
         Seconds calculationInterval { 1_s };
         MonotonicTime lastCalculationTimestamp;
         unsigned frameCountSinceLastCalculation { 0 };
-        std::atomic<std::optional<double>> fps;
+        std::atomic<std::optional<float>> fps;
     } m_fpsCounter;
 
 #if ENABLE(DAMAGE_TRACKING)
